added simple example
added simple example
